requirements.txt
google-adk
google-genai
python-dotenv
psycopg2-binary
requests
slack-sdk
openai